Add article edit flow and use decimal prices
Add support for editing articles: new ArticoloModificaArticolo view model and ModificaArticolo POST action in ArticoliAdminController (handles name, description, price, category and image upload). Switch price types to decimal (AggiungiNuovoArticoloAdminViewModel and data-layer conversions) to improve monetary precision. Refactor GestioneDati: adjust SQL queries and readers, convert prezzo_listino to decimal when mapping Articolo, and tidy up multiple DB methods (Add/Update/Delete, associations, recommendations). Also minor fixes: cast cart sum in OrdiniController to double, remove IDE SQL dialect config, and add Deliveroo.csproj.DotSettings.user; several view and obj files updated accordingly.
IL PREZZO DELL'ARTICOLO DA DOUBLE E' UN DECIMAL DAPPERTUTTO (NON CAMBIA NIENTE PER IL DB MA MI VA MEGLIO A ME PER FARE I CONTI)

+++ Deliveroo/AdminViewModels/AggiungiNuovoArticoloAdminViewModel.cs
@@ -4,7 +4,7 @@ public class AggiungiNuovoArticoloAdminViewModel
 {
     public string Nome { get; set; }
     public string Descrizione { get; set; }
-    public double Prezzo { get; set; }
+    public decimal Prezzo { get; set; }
     public int IdCategoria { get; set; }
     public IFormFile UrlFoto { get; set; }
 }
