+++ Deliveroo/Tabelle/Articolo.cs
@@ -5,10 +5,8 @@ public class Articolo
     public int IdArticolo { get; set; }
     public string Nome { get; set; }
     public string Descrizione { get; set; }
-    public double Prezzo { get; set; }
+    public decimal Prezzo { get; set; }
     public int NumeroOrdini { get; set; }
     public Categoria Categoria { get; set; }
-    
-    //TEMPORANEO
     public string ImageUrl { get; set; }
 }
